Recurring Tasks (#4)
Adds initial support for recurring Tasks.

Recurring Tasks are Tasks that can be defined to automatically enqueue according to a given schedule expression.

Currently, the standard Cron parser from [github.com/robfig/cron](github.com/robfig/cron) is used and so it supports basic macros.

The current implementation involves having a given leader server poll for changes, and (re-using the existing leader lock mechanism) depends on mark-as-dirty and soft-deletes of RecurringTasks.  The periodic poll for changed-but-unseen data *and* a secondary polling sync-enforcer (check against all the recurring tasks defined in the data store) ensures that changes to recurring Tasks (creates, deletes, or updates), are taken into account. The polling intervals are configurable.

Each Task enqueued from a Recurring Task also has a `recurring_task_id` for tracking.

At the moment, there is no support for "do not enqueue if there is a previous un-handled Task" for Recurring Tasks.