[Config] Support env variables
Apparently Viper supports env variables unmarshaling after all, just not when the config
file does not contain the actual keys.

This PR updates the Dockerfile to use the default example config (which has all the keys)
as the file to put into `/app/config`, and updates the k8s deployment to use env variables
and secrets.

Signed-off-by: lloydmeta <lloydmeta@gmail.com>