[Archival] Add support for archiving completed Tasks
This adds a method to tasks.Service to find completed (DEAD or DONE) Tasks
that were completed before a given time and delete time from their current
index and add them to an archived index. This archive index can be rolled over
using ILM if desired.

It then adds a recurring function that calls this according to a configurable
duration.

Other added things:
* Index template for the archive index
* Kibana Index Pattern for the archive index

Signed-off-by: lloydmeta <lloydmeta@gmail.com>